[Refacto] Change company id type from int to long

+++ src/main/java/com/excilys/formation/cdb/mapper/ComputerMapper.java
@@ -23,7 +23,7 @@ public enum ComputerMapper {
 		c.setName(result.getString("name"));
 		c.setIntroduced(ldIntro);
 		c.setDiscontinued(ldDiscont);
-		c.setCompany(new Company(result.getInt("company_id"), result.getString("company_name")));
+		c.setCompany(new Company(result.getLong("company_id"), result.getString("company_name")));
 		
 		return c;
 	}
